add: Planned, coded, and tested the memory reallocator.

+++ include/safe_util/safe_manage.h
@@ -24,7 +24,7 @@
 
 typedef struct safe_manage {
     void *ptr;
-    size_t allocated;
+    long allocated;
 } safe_manage;
